# Flatten Bitmap formats_map into an array