Put most common formats first and do a linear scan.

Shown to be faster than map, unordered_map, and
lower_bound on sorted array.

I have a google bench setup in another branch which is part of the message PR queue. I wanted to present this one earlier since it's standalone.

```
static void BM_FindFormatSingle(benchmark::State& state) {
    const auto fmt = format_R8G8B8A8_a().format();
    for (auto _: state) {
        BitmapAccess::find_format(fmt);
    }
}

BENCHMARK(BM_FindFormatSingle);

static void BM_FindFormat(benchmark::State& state) {
    for (auto _: state) {
        BitmapAccess::find_format(format_R8G8B8A8_a().format());
        BitmapAccess::find_format(format_B8G8R8A8_a().format());
        BitmapAccess::find_format(format_A8B8G8R8_n().format());
        BitmapAccess::find_format(format_B8G8R8A8_n().format());
    }
}

BENCHMARK(BM_FindFormat);
```

Master:
```
BM_FindFormatSingle         19 ns         18 ns   38761196
BM_FindFormat               59 ns         58 ns   11620732
```

This PR:
```
BM_FindFormatSingle          3 ns          3 ns  227203497
BM_FindFormat               13 ns         13 ns   54225434
```