@fmatthew5876 2019-12-01
Not sure it's safe to remove the lazy init now until we can have #1960 

So I put it back in. Results are still very good:
```
BM_FindFormatSingle          5 ns          5 ns  121052213                                                                                                                   
BM_FindFormat               18 ns         18 ns   37775052
```